expose aura data provider from the client
