+++ template/node/Cargo.toml
@@ -67,6 +67,7 @@ pallet-transaction-payment = { workspace = true }
 
 # Frontier
 fc-api = { workspace = true }
+fc-aura = { workspace = true }
 fc-cli = { workspace = true }
 fc-consensus = { workspace = true }
 fc-db = { workspace = true }
